API endpoints logic added for all entities controller

+++ root/backend/src/main/java/com/ecommerce/controller/AdminPhotoController.java
@@ -1,5 +1,7 @@
 package com.ecommerce.controller;
 
+
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +16,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ecommerce.dto.CreatePhotoRequest;
 import com.ecommerce.entities.Photo;
+import com.ecommerce.entities.Role;
+import com.ecommerce.entities.User;
 import com.ecommerce.response.MessageResponse;
 import com.ecommerce.services.PhotoService;
 import com.ecommerce.services.UserService;
@@ -21,44 +25,72 @@ import com.ecommerce.services.UserService;
 @RestController
 @RequestMapping("/api/admin/photo")
 public class AdminPhotoController {
-	
-	@Autowired
-	private PhotoService photoService;
-	
 
-	@Autowired
-	private UserService userService;
-	
-	@PostMapping
-	public ResponseEntity<Photo> createPhoto(@RequestBody CreatePhotoRequest req, 
-											 @RequestHeader("Authorization") String jwt)throws Exception{
-		
-//		User user= userService.findUserByJwtToken(jwt);
-		Photo photo = photoService.createPhoto(req, req.getPhotoCategory());
-		
-		return new ResponseEntity<>(photo, HttpStatus.CREATED);
-	}
-	
-	@DeleteMapping("/{id}")
-	public ResponseEntity<MessageResponse> deletePhoto(@PathVariable Long id, 
-											 @RequestHeader("Authorization") String jwt) throws Exception{
-		
-//		User user= userService.findUserByJwtToken(jwt);
-		photoService.deletePhoto(id);
-		
-		MessageResponse res= new MessageResponse();
-		res.setMessage("Photo deleted successfully");
-		return new ResponseEntity<>(res, HttpStatus.CREATED);
-	}
-	
-	@PutMapping("/{id}")
-	public ResponseEntity<Photo> updatePhotoAvailabilityStatus(@PathVariable Long id, 
-											 @RequestHeader("Authorization") String jwt) throws Exception{
-		
-//		User user= userService.findUserByJwtToken(jwt);
-		Photo photo = photoService.updatePhotoAvailabilityStatus(id);
-		
-		return new ResponseEntity<>(photo, HttpStatus.CREATED);
-	}
-	
+    @Autowired
+    private PhotoService photoService;
+
+    @Autowired
+    private UserService userService;
+
+    @PostMapping
+    public ResponseEntity<Photo> createPhoto(@RequestBody CreatePhotoRequest req,
+                                             @RequestHeader("Authorization") String jwt) throws Exception {
+
+        // Retrieve the user from the JWT token
+        User user = userService.findUserByJwtToken(jwt)
+                                .orElseThrow(() -> new Exception("User not authenticated"));
+
+        // Check if the user has ADMIN role
+        if (!user.getRole().equals(Role.ADMIN)) {
+            throw new Exception("User does not have permission to create photos");
+        }
+
+        // Create the photo
+        Photo photo = photoService.createPhoto(req, req.getPhotoCategory());
+
+        return new ResponseEntity<>(photo, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<MessageResponse> deletePhoto(@PathVariable Long id,
+                                                       @RequestHeader("Authorization") String jwt) throws Exception {
+
+        // Retrieve the user from the JWT token
+        User user = userService.findUserByJwtToken(jwt)
+                                .orElseThrow(() -> new Exception("User not authenticated"));
+
+        // Check if the user has ADMIN role
+        if (!user.getRole().equals(Role.ADMIN)) {
+            throw new Exception("User does not have permission to delete photos");
+        }
+
+        // Delete the photo
+        photoService.deletePhoto(id);
+
+        MessageResponse res = new MessageResponse();
+        res.setMessage("Photo deleted successfully");
+        return new ResponseEntity<>(res, HttpStatus.OK);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Photo> updatePhotoAvailabilityStatus(@PathVariable Long id,
+                                                               @RequestHeader("Authorization") String jwt) throws Exception {
+
+        // Retrieve the user from the JWT token
+        User user = userService.findUserByJwtToken(jwt)
+                                .orElseThrow(() -> new Exception("User not authenticated"));
+
+        // Check if the user has ADMIN role
+        if (!user.getRole().equals(Role.ADMIN)) {
+            throw new Exception("User does not have permission to update photo availability status");
+        }
+
+        // Update the photo availability status
+        Photo photo = photoService.updatePhotoAvailabilityStatus(id);
+
+        return new ResponseEntity<>(photo, HttpStatus.OK);
+    }
+
+    // Other endpoints...
 }
+
